Remove explicit sel_ldr_x86_* handling.
`go test` and `go run` already do this automatically for `GOOS=nacl`,
but only if you have the helper scripts in $PATH. If not, `go` just
skips the helper, with slightly confusing results.

However, using the Go-specific helpers has a distinct advantage: it
stops spamming stderr with messages like

    [29118,2110486400:09:30:00.034868] Native Client module will be loaded at base address 0x0000424500000000
    runtime: nacl_exception_stack: error 38
    runtime: nacl_exception_handler: error 38
    runtime: nacl_exception_stack: error 38
    runtime: nacl_exception_handler: error 38
    runtime: nacl_exception_stack: error 38
    runtime: nacl_exception_handler: error 38
    runtime: nacl_exception_stack: error 38
    runtime: nacl_exception_handler: error 38

So `go-nacl` is switching to using these helpers. Install them with
something like this (assuming `~/bin` is in `$PATH`):

    ln -s "$(go env GOROOT)"/misc/nacl/go_nacl_* ~/bin/

The remaining value of this wrapper is

1. easier to type than `GOOS=nacl go`
2. automatically maps `GOARCH=amd64` to `amd64p32`